make basic page work

+++ common-lib/src/main/kotlin/xyz/staffjoy/common/utils/MD5Util.kt
@@ -3,14 +3,16 @@ package xyz.staffjoy.common.utils
 import java.io.UnsupportedEncodingException
 import java.security.MessageDigest
 import java.security.NoSuchAlgorithmException
+import kotlin.experimental.and
+import kotlin.experimental.or
 
 // helper from https://en.gravatar.com/site/implement/images/java/
 object MD5Util {
     fun hex(array: ByteArray): String {
         val sb = StringBuffer()
         for (i in array.indices) {
-            sb.append(Integer.toHexString((array[i]
-                    and 0xFF) or 0x100).substring(1, 3))
+            sb.append(Integer.toHexString(((array[i]
+                    and 0xFF.toByte()) or 0x100.toByte()).toInt()).substring(1, 3))
         }
         return sb.toString()
     }
